requirements.txt
appdirs==1.4.4
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cssselect==1.2.0
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
exceptiongroup==1.0.4
fake-useragent==1.1.1
flake8==6.0.0
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
gunicorn==20.1.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
lxml==4.9.1
mccabe==0.7.0
oauthlib==3.2.2
outcome==1.2.0
parse==1.19.0
protobuf==4.21.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyee==8.2.2
pyflakes==3.0.1
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
pytz==2022.2.1
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
w3lib==2.1.0
websockets==10.4
wsproto==1.2.0
zipp==3.11.0
